Moved as_finite_diff to calculus.finite_diff (no hook in core).
