+++ sympy/calculus/__init__.py
@@ -3,4 +3,4 @@ SymPy modules tree.
 """
 
 from .singularities import singularities
-from .finite_diff import finite_diff_weights, apply_finite_diff
+from .finite_diff import finite_diff_weights, apply_finite_diff, as_finite_diff
