+++ sympy/calculus/finite_diff.py
@@ -7,14 +7,17 @@ difference weights for ordinary differentials of functions for
 derivatives from 0 (interpolation) up to arbitrary order.
 
 The core algorithm is provided in the finite difference weight generating
-method (finite_diff_weights), and a convenience function for estimating
-a derivative (or interpolate) directly from a series of points
-is also provided (apply_finite_diff).
+function (finite_diff_weights), and two convenience functions are provided
+for:
+
+- estimating a derivative (or interpolate) directly from a series of points
+is also provided (``apply_finite_diff``).
+- making a finite difference approximation of a Derivative instance (``as_finite_diff``).
 
 """
 
 from sympy import S
-
+from sympy.core.compatibility import iterable
 
 def finite_diff_weights(order, x_list, x0):
     """
@@ -239,3 +242,111 @@ def apply_finite_diff(order, x_list, y_list, x0):
     for nu in range(0, len(x_list)):
         derivative += delta[order][N][nu]*y_list[nu]
     return derivative
+
+
+def as_finite_diff(derivative, points=1, x0=None, wrt=None):
+    """
+    Returns an approximation of a derivative of a function in
+    the form of a finite difference formula. The expression is a
+    weighted sum of the function at a number of discrete values of
+    (one of) the independent variable(s).
+
+    Parameters
+    ==========
+
+    derivative: a Derivative instance (needs to have an variables
+        and expr attribute).
+
+    points: sequence or coefficient, optional
+        If sequence: discrete values (length >= order+1) of the
+        independent variable used for generating the finite
+        difference weights.
+        If it is a coefficient, it will be used as the step-size
+        for generating an equidistant sequence of length order+1
+        centered around x0. defult: 1 (step-size 1)
+
+    x0: number or Symbol, optional
+        the value of the independent variable (wrt) at which the
+        derivative is to be approximated. default: same as wrt
+
+    wrt: Symbol, optional
+        "with respect to" the variable for which the (partial)
+        derivative is to be approximated for. If not provided it
+        is required that the Derivative is ordinary. default: None
+
+
+    Examples
+    ========
+
+    >>> from sympy import symbols, Function, exp, sqrt, Symbol, as_finite_diff
+    >>> x, h = symbols('x h')
+    >>> f = Function('f')
+    >>> as_finite_diff(f(x).diff(x))
+    -f(x - 1/2) + f(x + 1/2)
+
+    The default step size and number of points are 1 and ``order + 1``
+    respectively. We can change the step size by passing a symbol
+    as a parameter:
+
+    >>> as_finite_diff(f(x).diff(x), h)
+    -f(-h/2 + x)/h + f(h/2 + x)/h
+
+    We can also specify the discretized values to be used in a sequence:
+
+    >>> as_finite_diff(f(x).diff(x), [x, x+h, x+2*h])
+    -3*f(x)/(2*h) + 2*f(h + x)/h - f(2*h + x)/(2*h)
+
+    The algorithm is not restricted to use equidistant spacing, nor
+    do we need to make the approximation around x0, but we can get
+    an expression estimating the derivative at an offset:
+
+    >>> e, sq2 = exp(1), sqrt(2)
+    >>> xl = [x-h, x+h, x+e*h]
+    >>> as_finite_diff(f(x).diff(x, 1), xl, x+h*sq2)
+    2*h*((h + sqrt(2)*h)/(2*h) - (-sqrt(2)*h + h)/(2*h))*f(E*h + x)/\
+    ((-h + E*h)*(h + E*h)) + (-(-sqrt(2)*h + h)/(2*h) - \
+    (-sqrt(2)*h + E*h)/(2*h))*f(-h + x)/(h + E*h) + \
+    (-(h + sqrt(2)*h)/(2*h) + (-sqrt(2)*h + E*h)/(2*h))*f(h + x)/(-h + E*h)
+    
+    Partial derivatives are also supported:
+
+    >>> y = Symbol('y')
+    >>> d2fdxdy=f(x,y).diff(x,y)
+    >>> as_finite_diff(d2fdxdy, wrt=x)
+    -f(x - 1/2, y) + f(x + 1/2, y)
+
+    See also
+    ========
+
+    sympy.calculus.finite_diff.apply_finite_diff
+    sympy.calculus.finite_diff.finite_diff_weights
+
+    """
+    if wrt is None:
+        wrt = derivative.variables[0]
+        # we need Derivative to be univariate to guess wrt
+        if any(v != wrt for v in derivative.variables):
+            raise ValueError('if the function is not univariate'+
+                             ' then `wrt` must be given')
+
+    order = derivative.variables.count(wrt)
+
+    if x0 is None:
+        x0 = wrt
+
+    if not iterable(points):
+        # points is simply the step-size, let's make it a
+        # equidistant sequence centered around x0
+        if order % 2 == 0:
+            # even order => odd number of points, grid point included
+            points = [x0 + points*i for i
+                      in range(-order//2, order//2 + 1)]
+        else:
+            # odd order => even number of points, half-way wrt grid point
+            points = [x0 + points*i/S(2) for i
+                      in range(-order, order + 1, 2)]
+
+    if len(points) < order+1:
+        raise ValueError("Too few points for order %d" % order)
+    return apply_finite_diff(order, points, [
+        derivative.expr.subs({wrt: x}) for x in points], x0)
